conversions_is_OK

+++ exercises/conversions/using_as.rs
@@ -1,20 +1,18 @@
 // using_as.rs
 //
-// Type casting in Rust is done via the usage of the `as` operator. Please note
-// that the `as` operator is not only used when type casting. It also helps with
-// renaming imports.
+
+// 在 Rust 中，类型转换是通过使用 as 运算符完成的。请注意，as 运算符不仅用于类型转换，它还用于给导入项重命名。。
 //
-// The goal is to make sure that the division does not fail to compile and
-// returns the proper type.
-//
+// 目标是确保除法不会导致编译失败，并返回正确的类型。
 // Execute `rustlings hint using_as` or use the `hint` watch subcommand for a
 // hint.
 
-// I AM NOT DONE
+// I AM
+// DONE
 
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / (values.len() as f64)
 }
 
 fn main() {
